Made the allocator more robust
